+++ lib/adapters/idb/idb-bulk-docs.js
@@ -17,7 +17,7 @@ var decodeMetadata = idbUtils.decodeMetadata;
 var encodeMetadata = idbUtils.encodeMetadata;
 var idbError = idbUtils.idbError;
 
-function idbBulkDocs(req, opts, api, idb, Changes, callback) {
+function idbBulkDocs(req, opts, api, Changes, callback) {
   var docInfos = req.docs;
   var txn;
   var docStore;
@@ -60,7 +60,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
       ATTACH_STORE, META_STORE,
       LOCAL_STORE, ATTACH_AND_SEQ_STORE
     ];
-    txn = idb.transaction(stores, 'readwrite');
+    txn = api._idb.transaction(stores, 'readwrite');
     txn.onerror = idbError(callback);
     txn.ontimeout = idbError(callback);
     txn.oncomplete = complete;
