+++ lib/adapters/idb/idb-utils.js
@@ -229,3 +229,6 @@ exports.compactRevs = function (revs, docId, txn) {
     };
   });
 };
+
+exports.cachedDBs = new utils.Map();
+exports.openReqList = new utils.Map();
